+++ internal/tools/get_cross_space_context.go
@@ -100,11 +100,11 @@ func registerGetCrossSpaceContext(s *mcpserver.MCPServer, client *webex.Client) 
 			text += fmt.Sprintf("### %s (%d match%s)\n\n", sr.Title, len(sr.Matches), pluralize(len(sr.Matches)))
 			for _, m := range sr.Matches {
 				if m.Before != nil {
-					text += fmt.Sprintf("  _%s (%s)_: %s\n", m.Before.PersonEmail, m.Before.Created, truncate(m.Before.Text, 100))
+					text += fmt.Sprintf("  _%s (%s)_: %s\n", m.Before.PersonEmail, m.Before.Created, sandboxText(truncate(m.Before.Text, 100)))
 				}
-				text += fmt.Sprintf("  **%s (%s): %s** ← match\n", m.Message.PersonEmail, m.Message.Created, truncate(m.Message.Text, 150))
+				text += fmt.Sprintf("  **%s (%s): %s** ← match\n", m.Message.PersonEmail, m.Message.Created, sandboxText(truncate(m.Message.Text, 150)))
 				if m.After != nil {
-					text += fmt.Sprintf("  _%s (%s)_: %s\n", m.After.PersonEmail, m.After.Created, truncate(m.After.Text, 100))
+					text += fmt.Sprintf("  _%s (%s)_: %s\n", m.After.PersonEmail, m.After.Created, sandboxText(truncate(m.After.Text, 100)))
 				}
 				text += "\n"
 			}